requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.19.2
psycopg2-binary==2.9.3
django-split-settings==1.1.0
django-debug-toolbar==3.2.4
gunicorn==20.1.0
